Load 1.18 worlds
