+++ pizzaserver-server/src/main/java/io/github/pizzaserver/server/level/ImplLevelManager.java
@@ -6,9 +6,8 @@ import io.github.pizzaserver.api.level.providers.ProviderType;
 import io.github.pizzaserver.api.level.world.data.Dimension;
 import io.github.pizzaserver.commons.utils.ReadWriteKeyLock;
 import io.github.pizzaserver.format.api.BedrockLevel;
-import io.github.pizzaserver.format.api.chunks.BedrockChunk;
-import io.github.pizzaserver.format.api.chunks.BedrockChunkProvider;
-import io.github.pizzaserver.format.mcworld.MCWorldLevel;
+import io.github.pizzaserver.format.api.dimension.chunks.BedrockChunk;
+import io.github.pizzaserver.format.api.provider.mcworld.MCWorldLevel;
 import io.github.pizzaserver.server.ImplServer;
 import io.github.pizzaserver.server.level.processing.LevelChunkProcessorManager;
 import io.github.pizzaserver.server.level.world.ImplWorld;
@@ -112,7 +111,7 @@ public class ImplLevelManager implements LevelManager, Closeable {
         if (!file.exists()) {
             throw new FileNotFoundException("No level exists with the name: " + name);
         }
-        BedrockLevel<? extends BedrockChunkProvider<? extends BedrockChunk>> provider;
+        BedrockLevel provider;
         try {
             provider = this.getProvider(file, ProviderType.resolveByFile(file));
         } catch (IOException exception) {
@@ -200,7 +199,7 @@ public class ImplLevelManager implements LevelManager, Closeable {
      * @return provider
      * @throws IOException if an exception occurred while reading the file
      */
-    protected BedrockLevel<? extends BedrockChunkProvider<? extends BedrockChunk>> getProvider(File levelFile, ProviderType providerType) throws IOException {
+    protected BedrockLevel getProvider(File levelFile, ProviderType providerType) throws IOException {
         switch (providerType) {
             case LEVELDB:
                 return new MCWorldLevel(levelFile);
