+++ pizzaserver-server/src/main/java/io/github/pizzaserver/server/player/ImplPlayer.java
@@ -51,6 +51,7 @@ import io.github.pizzaserver.server.entity.inventory.ImplPlayerInventory;
 import io.github.pizzaserver.server.level.world.ImplWorld;
 import io.github.pizzaserver.server.network.data.LoginData;
 import io.github.pizzaserver.server.network.protocol.PlayerSession;
+import io.github.pizzaserver.server.network.protocol.ServerProtocol;
 import io.github.pizzaserver.server.player.handlers.AuthInputHandler;
 import io.github.pizzaserver.server.player.handlers.InventoryTransactionHandler;
 import io.github.pizzaserver.server.player.handlers.PlayerPacketHandler;
@@ -218,7 +219,7 @@ public class ImplPlayer extends ImplEntityHuman implements Player {
             startGamePacket.setGeneratorId(1);
             startGamePacket.setDefaultPlayerPermission(PlayerPermission.MEMBER);
             startGamePacket.setServerChunkTickRange(this.getServer().getConfig().getChunkRadius());
-            startGamePacket.setVanillaVersion("1.17.40");
+            startGamePacket.setVanillaVersion(ServerProtocol.LATEST_GAME_VERSION);
             startGamePacket.setPremiumWorldTemplateId("");
             startGamePacket.setInventoriesServerAuthoritative(true);
             startGamePacket.getExperiments().add(new ExperimentData("data_driven_items", true));
