+++ pizzaserver-worldfile/src/main/java/io/github/pizzaserver/format/api/chunks/ChunkParseException.java
@@ -1,0 +1,15 @@
+package io.github.pizzaserver.format.api.chunks;
+
+import java.io.IOException;
+
+public class ChunkParseException extends IOException {
+
+    public ChunkParseException(String message) {
+        super(message);
+    }
+
+    public ChunkParseException(String message, Throwable throwable) {
+        super(message, throwable);
+    }
+
+}
