+++ pizzaserver-worldfile/src/main/java/io/github/pizzaserver/format/api/data/DimensionIds.java
@@ -1,0 +1,12 @@
+package io.github.pizzaserver.format.api.data;
+
+public class DimensionIds {
+
+    public static final int OVERWORLD = 0;
+    public static final int NETHER = 1;
+    public static final int END = 2;
+
+
+    private DimensionIds() {}
+
+}
