+++ pizzaserver-worldfile/src/main/java/io/github/pizzaserver/format/api/dimension/chunks/BedrockBiomeMap.java
@@ -1,0 +1,37 @@
+package io.github.pizzaserver.format.api.dimension.chunks;
+
+import io.github.pizzaserver.commons.utils.Check;
+import io.github.pizzaserver.format.api.dimension.chunks.subchunk.BedrockSubChunkBiomeMap;
+import io.github.pizzaserver.format.api.dimension.chunks.subchunk.utils.Palette;
+
+import java.util.*;
+
+public class BedrockBiomeMap {
+
+    private final Map<Integer, BedrockSubChunkBiomeMap> subChunkBiomes = new HashMap<>();
+
+
+    /**
+     * Retrieve a sub chunk biome map if one exists. Otherwise return null
+     * @param subChunk sub chunk index
+     * @return the biome map at that index if one exists
+     */
+    public BedrockSubChunkBiomeMap getSubChunk(int subChunk) {
+        this.subChunkBiomes.computeIfAbsent(subChunk, ignored -> new BedrockSubChunkBiomeMap(new Palette<>()));
+        return this.subChunkBiomes.get(subChunk);
+    }
+
+    /**
+     * Set a sub chunk biome map at an index.
+     * @param subChunk sub chunk index
+     * @param biomeMap the biome map at that index
+     */
+    public void setSubChunk(int subChunk, BedrockSubChunkBiomeMap biomeMap) {
+        this.subChunkBiomes.put(subChunk, biomeMap);
+    }
+
+    public List<BedrockSubChunkBiomeMap> getSubChunks() {
+        return new ArrayList<>(this.subChunkBiomes.values());
+    }
+
+}
