--- pizzaserver-worldfile/src/main/java/io/github/pizzaserver/format/exceptions/world/chunks/ChunkParseException.java
@@ -1,15 +1,0 @@
-package io.github.pizzaserver.format.exceptions.world.chunks;
-
-import java.io.IOException;
-
-public class ChunkParseException extends IOException {
-
-    public ChunkParseException(String message) {
-        super(message);
-    }
-
-    public ChunkParseException(String message, Throwable throwable) {
-        super(message, throwable);
-    }
-
-}
