--- pizzaserver-worldfile/src/main/java/io/github/pizzaserver/format/exceptions/world/chunks/ChunkSerializationException.java
@@ -1,13 +1,0 @@
-package io.github.pizzaserver.format.exceptions.world.chunks;
-
-public class ChunkSerializationException extends RuntimeException {
-
-    public ChunkSerializationException(String message) {
-        super(message);
-    }
-
-    public ChunkSerializationException(String message, Throwable throwable) {
-        super(message, throwable);
-    }
-
-}
